Consolidate deployed-agent config into one Configurations and Secrets drawer
